+++ src/pytorch_lightning/loggers/wandb.py
@@ -31,9 +31,10 @@ from pytorch_lightning.utilities.logger import _add_prefix, _convert_params, _fl
 from pytorch_lightning.utilities.rank_zero import rank_zero_only, rank_zero_warn
 
 try:
-    import wandb
     from wandb.sdk.lib import RunDisabled
     from wandb.wandb_run import Run
+
+    import wandb
 except ModuleNotFoundError:
     # needed for test mocks, these tests shall be updated
     wandb, Run, RunDisabled = None, None, None  # type: ignore
@@ -223,7 +224,7 @@ class WandbLogger(Logger):
 
     Args:
         name: Display name for the run.
-        save_dir: Path where data is saved (wandb dir by default).
+        save_dir: Path where data is saved.
         offline: Run offline (data can be streamed later to wandb servers).
         id: Sets the version, mainly used to resume a previous run.
         version: Same as id.
@@ -255,7 +256,7 @@ class WandbLogger(Logger):
     def __init__(
         self,
         name: Optional[str] = None,
-        save_dir: Optional[str] = None,
+        save_dir: str = ".",
         offline: bool = False,
         id: Optional[str] = None,
         anonymous: Optional[bool] = None,
@@ -300,7 +301,7 @@ class WandbLogger(Logger):
             name=name,
             project=project,
             id=version or id,
-            dir=save_dir,
+            dir=save_dir or kwargs.pop("dir"),
             resume="allow",
             anonymous=("allow" if anonymous else None),
         )